+++ WebContent/WEB-INF/commandHandlerURI.properties
@@ -6,10 +6,8 @@
 /changePwd.do=auth.controller.ChangePwdController
 
 # 관리자 페이지
-/admin/memberList.do=admin.controller.ListMemberController
-/admin/updateGrade.do=admin.controller.MemberInfoController
-/admin/boardList.do=admin.controller.ListBoardController
-/admin/deleteBoard.do=admin.controller.DeleteBoardController
+/admin/memberList.do=admin.controller.MemberListController
+/admin/boardList.do=admin.controller.BoardListController
 
 # 메인 페이지
 /main.do=main.controller.MainPageController
